Minor improvements to rust code
