+++ crates/node-bindings/src/core/filesystem/in_memory_file_system/mod.rs
@@ -1,5 +1,3 @@
-//! Utilities for testing parcel core filesystem
-
 use parcel_resolver::FileSystem;
 use std::collections::HashMap;
 use std::path::{Component, Path, PathBuf};
